[guiconnect] - simple locking with a mutex
See my extended comment on the Yoshimi-developers mailing list
"Yoshimi GUI connection(WIP)" from 22.1.2024

As can be seen, synchronisation with a mutex is very simple and clear
and easy to get correct. Since both the writer and the reader pass through
a barrier, we can ensure that a writer can not corrupt the data of an older
slot which is concurrently read by the GUI, and we ensure that the GUI can not
see stale data, since the CPU cache is published to memory whenever leaving
a critical section.

The same effect could be achieved with std::atomic.
Yet doing so is much more tricky (both sides would have to loop and
retry when detecting a concurrent modification). And the relevance
of the performance gain remains questionable (at least until we can
judge and observe see the actual usage of this system in the code base).