[guiconnect] - wire new connections for the Effect UI
This is a first attempt to introduce GuiDataExchange connections into UI components.
Unfortunately FLTK requires a fixed constructor signature for widgets;
thus we're forced to allow for (temporarily) uninitialised connections;
however, the plan is to have an init() call immediately after creating
the window or widget, so that we can effectively always assume an active connection.

So the wiring pattern introduced with this change set works as follows:

- in the SynthEngine (or some embedded component), for each GuiDataExchange "channel"
  a distinct connection object is created
- when the GUI starts up, SynthEngine::publishGuiAnchor() extracts
  the corresponding RoutingTags and places them into the InterfaceAnchor record,
  which is then sent per push-update into the GUI
- MasterUI creates the windows and invokes init()
- the EffUI::init() retrieves the suitable RuitingTag, combines them
  with a reference to GuiDataExchange and activates the MirrorData elements

Result:
- each EffUI instance has now a connected MirrorData<EffectDTO> in the field "effect"
- and the embedded EQGrap instance has likewise a MirrorData<EqGraphDTO> in the field "graph"

The next planned step will be to trigger a push-update on each relevant Effect/EQ change
After that we can start to switch from direct Core data access to using the
data in the MirrorData instances