[guiconnect] - identify situations where to trigger a push update
After exhaustive analysis of the update mechanisms
it became clear that we have to care for two different situations

- a change of the currently exposed component in the GUI
- intrinsic data changes for effect parameters

It seems to fit best with the existing style of handling changes to attach
the push updates to the *command processing* (as opposed to attach it to
actual data changes). All relevant manipulations, either through UI widgets
or from the CLI will propagate to
 - InterChange::commandSysIns
 - InterChange::commandEffects

Using this approach, we have to care for some additional situations where
the Effect state is changes as result of global actions, like loading state.
Analysis of call graphs inidcates that these all boil down to loading data
from XML (with a sole exception, the SynthEngine::resetAll() function).
Moreover, no update is required from Init(), since the default state of the
GUI will be propagated correctly from a default constructed EffectDTO.