ensure defaults() before loading XML
This change rectifies usage of SynthEngine::defaults() to ensure
loading of settings from XML is preceded in all cases by a global reset.
SynthEngine::getfromXML() is called from
 - SynthEngine::loadXML(filename)
 - SynthEngine::putalldata(*data, size)
 - Config::restoreSessionData(sessionfile)

Asymmetry in calls was discovered during analysis for the GuiConnect refactoring;
as it turned out, we missed that defaults() call when loading a JackSession.

SynthEngine::defaults() itself was checked to be idempotent