resolve possible memory leaks using XMLwrapper by switching to unique_ptr
The ZynAddSubFX codebase was originally written in some "C with objects" dialect;
thus heap allocations are still handled in C-style at various places in the code,
which requires to delete each allocation explicitly on each possible code path.
This falls short for several reasons:
- in code with elaborate control logic it is very hard really to cover all execution paths
- whenever any invoked function throws an exception, we'd have to catch that explicitly
  to do the local clean-up and then re-throw the exception

*note* other than in C, a heap allocation (operator new) can never return NULL in C++
Rather, it will throw a std::bad_alloc exception. Many XMLwrapper usages were written
under the assumption that an allocation failure can be detected locally by testing
for a NULL pointer.

The simple (and idiomatic C++) solution is to use an std::unique_ptr instead,
which will automatically delete the heap allocation whenever this smart-ptr object
goes out of scope, irrespective if this is due to regular execution or to an exception.

Moreover, this allows to get rid of pointers (which can be NULL and can be re-bound)
and use references in all functions which just /use/ an XMLwrapper.

Note: The only point where a NULL pointer was actually used to encode meaning
was in Bank::parseBanksFile(). Since we just recently switched to C++17.
this can now be better expressed using a std::optional to indicate the
special case that no bankFile is passed (and thus a new root shall be established)