[guiconnect] - extend push-update to handle part-effects as well
As it turns out, from within the component, part effects and
insert effects are indistinguishable, they are just initialised
by another Effect instance. The only tangible difference is the
handling of the selector and bypass widgets, which are not located
in the EffUI, but in the PartUI rather, and propagate to commandPart.

The code uses the partNr do encode the difference, making system effect
and global insert effect two special cases. Thus, following the same logic,
the push-update handler can be made generic and cover all three cases.

As an aside, this change fixes duplicate creation of PartUI

Seemingly related to
6c7a3061d

Yet in fact, PartUI will be automatically constructed by FLTK,
and thus the further initialisation code can be moved down to MasterUI::make_window().

Moreover, seemingly it is not necessary to add it manually to the partuigroup,
since this happens automatically when widgets are created before partuigroup->end()
Only the call to checkEngines() must happen after closing this ownership backet
and after the end of make_windows()