[guiconnect] - transfer effect data to DTO
Add the ability to the Effect interface to extract all parameters
by an bulk operation, thereby allowing to populate the DTO in one shot.
Most effect use <20 parameters, but the EQ is an "outlier", since it
exposes 5 settings per bank, making for a total of 45 parameters.

Used this as opportunity to introduce "override" markers into this
hierarchy; this is a good practice to prevent from accidentally
"detaching" some subclass virtual function from the interface.