[guiconnect] - switch complete EffUI from direct-access to using DTO data
This changeset even removes the EffMgr* from the UI.
Moreover, several abbreviations were introduced, to make the code more legible

- isBeatSync()
- isSepDelay()
- isInsert()
- effParam(n)
- effPreset()
- __setColor(widget, preset, idx)

NOTE: this is only a partial change, since other parts of the UI
still continue to access the core directly. This causes inconsistent
behaviour esp. in the Part-Effect, which will be addressed as next steps.
Moreover, the nested sub-displays are not yet adapted / switched to the
new system, and thus a "backdoor" is used to sneak in the direct pointer
into these sub components (-> DTO-field "eff_in_core_TODO_deprecated")