[guiconnect] - remove UI state regarding effect number and type
These UI state variables were used to perform an indexed access
directly into the Synth engine core. By removing those extraneous state
variables, the UI is no longer able to access the effect modules in the Core.

Rather, all usages were switched over to rely solely on the local values
sent per push-update and stored in the MirrorData block in the EffUI.
Notably this also removes the possiblility of core and UI state to get
mixed or confused; even after a command has been sent, the UI continues
to use the old values uniformely for all settings and parameters of that effect,
until the push-update arrives.

NOTE: A special twist with this changed scheme for UI connections
is that the core must now reliably issue a push-update whenever the
UI actually needs new or updated data. Failing to do so will leave
the UI with outdated state, since the UI does not actively pull data
from core any more.

Typically these updates should be attached to command processing,
because, as it stands, this is the most systematic way to know when
actually something happens or is changed in Yoshimi.

As an aside, I clarified the entrance point to the Unified Presets
(after I spent an hour searching where the heck the copy / paste
functionality is actually implemented; the old name "section()"
did not convey that it is actually a mutating operation)