[guiconnect] - a hack to force updates after load-by-config
While the previous changes provided fairly complete coverage of update situations
relevant to the effects, it left an open gap when changes happen /before GUI startup/

Notably this is the case when loading content per configuration or commandline
-l, --load=<file>
-L, --load-instrument=<file>[@part]
-S, --state[=<file>]

Even when push-updates are published in those cases, those messages will
not be delivered, rather silently discarded, because there is no receiver(yet).

So actually we'd need a "post-GUI-startup hook";
the SynthEngine could then decide if parts of the GUI need additional data
and (re)send some push-updates.

As a first and rather cheesy workaround, this changeset adds such a hook,
but invokes it directly from the GUI thread (it should better be run from Synth)
And as a second, complimentary hack, the GUI is prompted to do a refresh,
in order to make the received data visible. This is achieved by sending
the MAIN::control::masterReset, but *only* to the GUI.