[guiconnect] - switch EQ graph to a LUT delivered by push-update
Note: the old version used to call into the core for each pixel
with a frequency matched exactly to this pixel; with this changeset,
rather the draw() code uses local data from a table, which has the
downside that we have to interpolate between adjacent prepared
sample values. For very sharp peak/notch filters this can lead to
an unprecise, "wobbly" curve. The density of the LUT can be controlled
by the constant EQ_GRAPH_STEPS in Effect.h -- but beware: we send a float
for each point; so huge number of points ==> huge data block ==> an even
larger memory buffer must then be set aside for all the »slots« used by
GuiDataExchange. So this clearly is a trade-off