[guiconnect] - adapt the LV2 plugin to employ the new UI-start scheme
This solution works, but is not entirely correct,
since we generate a new UI-Anchor directly from the UI-Thread / Plugin-UI processing.
Thus we are missing a thread sync barrier towards the Synth / Plugin-Audio thread

In fact we do not know in which Thread the Synth-Object was created and
initialised by the LV2 plugin host. Actually, SynthEngine::publishGuiAnchor()
just generates a DTO and copies in the connection-IDs from the Synth-Object.

If this Synth-Object was created in another thread, there is the theoretical
possibility that some of these IDs were not yet published out of this thead's
CPU cache. While I should add, that in practice this seems highly unlikely.