Instances: use primary config as baseline for new instances
When a new Engine instance is created, the standard mechanism for loading config is invoked.
If the given instance number is used for the first time on the user's system,
this mechanism fails and gives a warning. This implicitly made fall all new instances
back on config default values, which are sane but quite different from what you'd expect
as a user, after having tweaked those values for "the application". Notably
some settings can not even be edited through the GUI.

An especially nasty example is the XML compression, which makes it effectively
impossible to fix the config by hand -- unless you're capable of Unix-Fu.

As a quick fix for this nuisance hereby a dedicated handler is introduced
to copy those settings which should be taken alongside for each pristine instance.

NOTE: the *actual* problem is that our Config object is an blatantly unorganised mess!
We should REALLY take the effort to sort it apart into relevant aspects.