Instances: change the instance registry to be keyed by ID
historically, the instance map was keyed by SynthEngine memory location,
which allowed to fetch the MusicClient for a given synth*. For some time,
this was used a lot, but most of those usages were superseded.

Today, only the CmdInterpreter needs to find several Synth instances,
and this can easily be accommodated to work per ID. This allows to change
the internal registry into a tree map, and to be thus sorted on IDs,
which in turn simplifies to allocate and maintain unique IDs.