Instances: SplashScreen component
Extract code to show a FLTK splash window.
The existing code used to construct FLTK widgets directly in the stack frame
in the main thread; doing so happens to work, due to destructor ordering, but is poentially
dangerous, because by convention FLTK adds follow-up widgets to a top-level window, which
takes owenership and deallocates those windows when destroyed.

The obvious solution is to encapsulate this GUI logic into a component,
allowing to heap-allocate all widgets and discard resources once the Splash
pop-up is no longer needed. Furthermore, we can utilise the timeout-callback
of FLTK for re-invocation.

These changes together help to simplify the main() code
and allow to get rid of FLTK includes at that level.