Refactoring: InterChange
This is the communication hub and as such always tied to one specific
SynthEngine instance. The InterChange can be made non-copyable, and
access its associated SynthEngine by reference.

Furthermore, the CommandBlock to be processed by the InterChange
always resides at a fixed memory location (after being retrieved from
some ringbuffer) and can thus be passed further down into the callgraph
by reference, thereby removing the possibility to re-bind it or pass NULL.

Initially, the code made an attempt to distinguish between "getData" == read commands
and "putData" == mutating commands. However, as the code evolved, seemingly this
distinction was transgressed at various places. Huge swaths of code now take
just a command block and figure out based on its contents if it is a read
or a mutating command. Thus it seems indicated to remove the possibly
misleading distinction in the variable names and name the reference
to the CommandData block simply as "cmd"