CMake: handle assertions properly; new option to activate diagnostics
The way the Yoshimi CMake build is organised will cause the default setup
for DEBUG or RELEASE builds to be completely overwritten with our flags.

This had the unfortunate effect that we missed NDEBUG on Release builds.
Contrary to common expectation we thus built releases with assertions enabled.

Sometimes you want assertions and debug information on release builds though.
A new Toggle "BuildForDiagnostic" will now address both concerns. By default,
it is *not* set and Release builds will then be built without debug info and
with "NDEBUG" set, causing all assertions to be disabled.
When *toggled on* however, even Release builds will be built with debug info
and without the "NDEBUG" flag.

Note: some setup options were renamed, to respect the grouping feature of CMake,
which is relevant when using some cmake-gui