Bugfix-LV2: start-up synchronisation Yoshimi-UI
with the GuiConnect branch, a synchronisation was introduced to
ensure the Yoshimi-LV2-core plugin is loaded completely before the
GUI-plugin will attempt to attach to the running SynthEngine.

The isReady-Flag was set in the activate() callback, which turned out
to be too late for the Reaper-DAW; this fix will set the isReady flag
a the point when the SynthEngine was booted and the instantiate() callback
is about to return the LV2_Handle to the host.