Fix initialisation of EQ filter bands
This problem was detected while testing the push-updates for the
EQ filter response graph in the UI, but the underlying issue turned out
to be present since a long time:

Whenever an EQ band was used for the first time,
the internal filter parameters were not properly synchronised with
the InterpolatedValue components used to actually control the AnalogFilter
sub-objects used for the actual implementation of the EQ.

This inconsistency resulted in the EQ sounding neutral on first use,
after initially setting up an filter type. Only after the frequency control
was nudged, the actual filter coefficients got computed and the filtering started
to work as intended.

The fix is explicitly to re-sync the filter parameters whenever the filter type
of an EQ band is changed.