BugFix: Not clearing 'old' config status unless change made.
This is triggered when a pristine install creates a full set of instrument banks.
Doing so, takes long enough to trigger a timeout. This *provisional* fix extends
the timeout for the first few calls. Experienced users could work round this, but
the majority would be unable to.
A more appropriate solution will be implemented as soon as possible.