Rework UI start: trigger bootstrap message directly
In order to receive push-updates with changed core state,
the GUI needs an InterfaceAnchor record with all the connection IDs.

This was introduced with v2.3.3 in an effort for better separation of
GUI and SynthEngine core; if this new connection system works well,
then we'll be able to remove the direct data access from GUI into
the SynthEngine step by step.

However, the first version of this setup used the new push-mechanism itself
to initiate a "bootstrap" message; this turned out to be fragile, especially
when start-up and especially loading of bank information takes some time.
The reason is that there is a timeout to prevent the GUI to be flooded with
already obsoleted messages in case the UI event processing was temporarily.

With this changeset, rather a direct access from the GUI thread is used
for the initial setup of this communication data, but protected by a mutex
to ensure data visibility between the SynthEngine and GUI threads, which
is especially relevant for LV2 where it is up to the host in which threads
to run the core-Plugin and the GUI-plugin.