Increase LV2 compatibility by supporting "UI" base class methods.
The interfaces "showInterface" and "idleInterface" are meant as
fallbacks if none of the sub class methods work or are supported.
This enables the jalv standalone host to launch Yoshimi, which wasn't
possible before.

Note that support for the sub class extension "external-ui" is still
preferred, since without it Yoshimi has no way to tell the host that
the UI was closed. So we keep that as the main class with the "UI"
base class as a fallback.

Signed-off-by: Kristian Amlie <kristian@amlie.name>