Config-LV2: keep instance config for LV2 apart from standalone instances
This is implemented by changing the way how instance-config files are located.
- for standalone : use the Synth-ID as key (as before)
- for LV2 : use a fixed string "LV2"

This implies that now a LV2-plugin instance does no longer _alias_ to some
standalone instance, which just happened to use the same Synth-ID.
Rather, all LV2-plugin instances now use a single common instance-config.

Note: usually the LV2 host pulls state data from each plug-in and
later stores this state-data internally, in the Host-Session.
This way, typically each plugin-instance has its own, private
patch-state.

But the base config and instance config is relevant for
default values when creating a new instance, and it can be
edited through the Config-UI in Yoshimi.


This changeset rearranges the code how the config-location is established,
and extracts it into a new helper function Config::buildConfigLocation()

The actual logic is only altered at a single point: when the instance-ID is generated.

The bulk of the further changes just eliminated directly generated config-file locations
to reuse this central helper function and the Config member fields instead.


Note: we need to set the Config::isLV2 flag earlier, because it is now
evaluated already when populating the config. We set this flag now
already from InstanceManager::Instance::startUp(pluginCreator),
i.e. before the Plugin-Instance itself is constructed.