Config-LV2: use the same patch-state handling for LV2 and standalone
The state saving for LV2 was introduced with
98d8b39f7b9749

It used its own implementation, residing in the SynthEngine;
With the later rework of config handling in general, there are now
two implementations for the same functionality, and in fact their
behaviour differs: the LV2 instance does not retrieve/populate the
instance config part; this can lead to sublte and possibly dangerous
errors when restoring a LV2 session on top of a different instance config
or even in a new setup with default values.

With this change, both use cases (state file and LV2 state) will now
invoke handlers in the Config object, both delegating to the same
core implementation sequence.


As an aside, this patch also fixes the following
 - remove unused variables Config::stateChanged and SynthEngine::usingYoshiType
 - fix the handling of missing config; one possible scenario is that a base config
   is found, but no instance config file is present. In this case, the base file
   should always be loaded, and only the instance file needs to be regenerated