Fix corruption bug in LV2 program names.
Because the original LV2 descriptor has a classic C char pointer,
Yoshimi's LV2 implementation augments that struct with a C++ string,
and then makes the former char pointer point to it. The problem is
that this pointer is immediately invalidated on the next line because
the string is copied into a vector, and the original struct is
destroyed. Hence the pointer is not valid anymore.

Simply using `std::move()` is not enough, because `string` is allowed
to store the string inside itself if it fits there, which would
invalidate the pointer even on moves. The effect of this is that long
names work, but short names do not.

The best and safest way is therefore to define both a move and a copy
constructor for the derived struct, which updates the pointer on every
move and copy.

We still switch to doing a `std::move()` at the original call site,
but this is entirely for performance reasons, and does not affect
correctness.

Signed-off-by: Kristian Amlie <kristian@amlie.name>