Tighten adherence to LV2 spec by loading programs in place.
The `LV2_Programs_Interface` extension's `select_program` call
explicitly says that a program selection should happen in a way so
that it will have been loaded the next time `run()` is called,
assuming that `select_program` has returned by then. This means that
we can not load in the background as we do now. This may introduce
xruns if the host calls the function from a RT context. The spec
doesn't say anything specific about which thread to call it from, but
two things are sort of implicit:

  1. If the host depends on the program having changed before the next
     `run()`, then we have no choice anyway, xruns are always a
     possibility.

  2. If it doesn't depend on it, but just wants the change to happen
     "as soon as possible", then it should call it from a low prio
     context. It might not, obviously, but then the bug is in the
     host.

One very important reason why this fix is necessary is the following
scenario:

  1. `~/.config/yoshimi-LV2.instace` has `ignore_program_change` set
     to 0.

  2. The user loads a project with an LV2 instance, with state data
     where `ignore_program_change` is set to 1.

  3. First, the program is set by the DAW. This results in a patch
     being scheduled to load, because so far, only the config data
     from point 1 is active. The word "scheduled" is important here.

  4. Afterwards, the state is applied, which causes a different patch
     (contained in the state data) to be loaded.

  5. *Then* the scheduled load is performed, even though
     `ignore_program_change` is now set to 1, because that condition
     was checked at scheduling time, not now. This patch is the
     incorrect one, since it is the one in the state data we want.

By loading in place (at point 3), we avoid the problem above, in
addition to adhering more closely to the spec.

Signed-off-by: Kristian Amlie <kristian@amlie.name>