Implement Omni mode support for each part.
Omni means to respond to notes and messages from all channel, not just
one. It can be enabled either in the part settings, or by sending the
standard MIDI 124 and 125 CC messages. The latter can be disabled
completely in the config settings if desired.

Includes work by Will Godfrey on the UI part.

Signed-off-by: Kristian Amlie <kristian@amlie.name>