MXML: Start rework of XML handling code for MXML-4 compatibility
The breaking API and ABI changes introduced with MXML v4.x pose a challenge,
since Yoshimi must be able to support both versions of this widely used base library
for several years to come. This was taken as an opportunity to loosen the direct
connection between Yoshimi core logic and MXML data structures -- creating an
internal interface that can be implemented differently for each library variant.

The existing XMLwrapper will be dissected into several concerns
 * XMLStore represents a structured data repository backed by XML
 * XMLtree is a (lightweight) access point to some data subtree
 * XMLtree::Node is a marker for MXML data and serves as policy class
   within the implementation backend  to abstract from those parts
   of the library API affected by the change...

This changeset is the initial setup for an effort to accomplish this refactoring.
A copy of the existing implementation is set apart under a new name
and will undergo the transformation over the following changesets.