MXML: add a test setup to support analysis and refactoring
Since the Yoshimi source does not provide any unit tests,
understanding code and refactoring can be challenging.

For the GuiDataExchange I successfull used an one-time
(and throw-away) developer test function for this purpose.
Thus I'll follow the same recipe here:

 - add a test function into the sound test enviroment, accessible from CLI
 - sample, copy-and-paste some typical XML reading and writing code

The test can be launched with

    build/yoshimi -i --null <<<"set test exec"

__Remark__: this simple exercise uncovers some quality problems with the existing code:
When invoking `XMLwrapper::loadXMLfile()` followed by `XMLwrapper::getXMLdata()`
- additional spuriouis whitespace is added
- each invocation of getXMLdata injects an additional copy of the xmltype by side-effect
- the XMLwrapper::info field points into de-allocated memory, and the mere
  invocation of `getXMLdata()` leads to memory corruption or segfault

Thus adding additional guards and fixes to the code to be reworked.