MXML: implement simplified render and save
An XMLStore should *only* handle the XML-tree.

It is *not* it's job to inject any magic metadata,
since the content of the data tree falls within the responsibility
of the Yoshimi core logic; notably any markers regarding the "XML type",
or flags if some Synth component (ADD, SUB, PAD) was used,
should be maintained by the core logic, not the XMLStore.

As a result, the XML handling code becomes much simpler and clearer.