MXML: switch core code -- UnifiedPresets
handles storing of preset files,
and also saves as backend for clipboard.

On this occasion, some simplifications were made,
without changing the overall decision logic:

- immediately unpack the relevant fields of the CommandBlock
  into fields in UnifiedPresets (-> this saves some lines in each function)
- make these fields immutable (-> uncovers some corner cases, to be treated directly)
- investigate passing of the "value" / "human" field, document it better,
  and remove some unnecessary assignments, since the code path is clear...
- extract also a dedicated function for effects and synth-components;
  after that, the structure of accessXML() can be simplified significantly

Remarks:
UnifiedPresets are a "point-and-shot" object. There is only a single usage site,
where this object is constructed and handleStoreLoad() is invoked; after that,
the object is discarded (which allows for good optimisation)

The call path relevant for this topic is:
 handlestoreLoad() -> load() / save() -> accessXML()

Typically, an invocation figures out from the contents of the CommandBlock
which part and component to treat, and derives the ID of an associated preset.
It then reads/writes XML data and returns the resolved preset-ID.