MXML: reimplement version compatibility checks
The old XMLwrapper performed some version compatibility checks,
in a homegrown and somewhat confusing way, since the code for this
evaluation was scattered at various places and used side-effects
and transport through a chain of transient state flags.

This is an attempt to re-implement the logic in a separate function --
to the degree I was able to figure out what was intended to happen.

(1) Check if there is any document type we can handle
(2) Check if this is a Zyn-document type with a version deemed as compatible
(3) Check if we are loading config from an earlier Yoshimi version

In all cases, these checks will only log warn messages
and set some marker flags in the Config object.

__Note__: the actual extraction of XML data happens independently
from these checks and is guarded with safety checks and attempts
to be as much compatible as possible. Migration of older config
data is handled on the level of the individual data fields.

Changes:
- complete re-implementation of compatibility check
- add dedicated predicate Config::is_compatible(otherVer)
  to check compatibility in accordance with »semantic versioning«
- remove `fileCompatible` state flag from `SynthEngine`
  and instead use `incompatibleZynFile` in `Config`
- also re-implement logging of XML headers,
  now directly based on the metadata in XMLStore

+ Fix problem with auto-saving of Config for version migration:
  - base config is not saved
  - instance config contains empty bank-root definition

Solution: perform these "migration" steps later, and explicitly invoked
from Instance-Manager, at a point where the instance is fully configured
and banks / roots were already loaded.
Furthermore, never automatically save a config imported from a newer
version into an older one, since changes might not be forward-compatible

Zyn-Compatibility: the last compatible version is defined as v2.4.3,
this is also the version Yoshimi as Zyn version number when writing files.
This value is hard wired in Config.cpp as VER_ZYN_COMPAT