Implement MIDI Learn for envelope freemode points.
When you click on the point it will learn both axes for that point,
and then you can choose afterwards whether to keep both or delete one
of them.

Since the original "insert" block command modifies both axes, it was
necessary to create two new block commands which modify each axis
independently, since MIDI Learn cannot work with two axes at
once. When learning the first block command, it is automatically
translated into the two other block commands.

Also rephrased some data strings to make the labels more descriptive,
in particular replacing the word "increment" with "time", which is the
same word used for ADSR curves (the "time" between each step, just
like the "time" between each point). There is still a problem with the
string which is displayed when actually doing the learning. I hope to
solve that in an upcoming commit.

Signed-off-by: Kristian Amlie <kristian@amlie.name>