Make messages better for FreeMode points, esp. with MIDI Learn.
We do this by adding support for two-dimensional message types in
the withValue function, and use that for the Point messages
specifically to print both the X and Y value.

Also rename the axes to X and Y. I did have them named Time and
Amplitude at one point, but the problem is that for the envelopes
deepest in the patch settings (typically Modulation), the label
becomes too long to fit in the MIDI Learn window. And in addition
"Amplitude" is not really correct when dealing with Frequency
envelopes. So just use the neutral "X and Y", which solves both
problems.

Signed-off-by: Kristian Amlie <kristian@amlie.name>