Config: keep stored setting separate for enabling GUI / CLI
This is an attempt to mitigate some problems with consistency.
Generally speaking, we want the config settings in the running instance
always be the same as the persisted state; thus (since last year) we
immediately persist any change made in the Config-UI,
to avoid pending state.

But there is a catch: once the user disables the GUI,
this change will immediately sever the communication between
UI and core, but it will not terminate the GUI, leaving it
running in unclear state.

As Johannes Zmölnig noticed, this inconsistency even prevents
the user to fix the disabled state of the GUI using the GUI.

With this change, we keep the persisted setting in separate
variables Config::storedCli and Config::storedGui;
while not touching the actual flags for the running instance.
Next time Yoshimi starts up, it will re-read the persisted state.

Furthermore, this changeset also ties the visible setting in
GUI and CLI to the persisted state. Even when the current
instance uses different settings, caused by an override
from the commandline (-i -I -c -C), we will always show
the setting from the presisted config.