MXML: Analyse situation and fix handling of text nodes
- investigate all breakages as indicated by the build log posted by Jonannes Zmölnig
  https://github.com/Yoshimi/yoshimi/issues/239#issue-3444006998

- read through the migration guide and investigate each point mentioned there
  https://www.msweet.org/mxml/mxml.html#migrating-from-mini-xml-v3.x

Conclusion: the necessary adaptations are minimal

* we need to pass the options and callbacks differently -> use a wrapper function
* some constants are named better -> introduce the better names already for the old version

Furthermore, it became clear that the Yoshimi code did not handle the **text nodes**
as was intended by MXML — yet it does in a way that happens to turn out correct.

Explanation:
By default, MXML stores *whitespace delimited* individual data elements as "Nodes" in the XML.
The special XML-Tags are also translated into Nodes, so that the complete document
can be represented as a sequence of nodes.

However, when some node needs to retain embedded whitespace, it can be marked as 'opaque'
Otherwise, a text content node shall only hold data without whitespace. If it is *precede*
by whitespace, that whitespace shall be striped, and rather the 'whitespace' flag
shall be set on the Node.

What Yoshimi does instead...
 * it packs the text content nodes into MXML_TYPE_TEXT
 * it sets the whitespace-flag to 'false' (so that MXML does not add whitespace)
 * but it does embed all kinds of whitespace into those text node,
   thereby **breaking the contract**
 * but then it reads back all its text content nodes as MXML_TYPE_OPAQUE

This changeset fixes this simply, by creating the text content nodes as MXML_TYPE_OPAQUE
right from beginning. Because that is what we want:
<string>....here we can do what we want... . . </string>