MXML: reorganise adaptation layer XMLtree::Node
This refactoring does not change functionality,
yet helps to highlight the purpose of the adaptor:
- for the rest of the code, Node is opaque
- but for the XMLtree implementation,
  it is called to translate into the MXML schematics.

This helps to map the meaning of "elements" and "content" as used in Yoshimi
onto the concepts of MXML, which are slightly different. Some calls can be
passed-through, while others (like text content) must be translated.

The Node inherits from the Policy class, which configures our usage of MXML
and allows to level the differences between old and new (v4) MXML API.