MXML: build the XML document root
The old Yoshimi code used a somewhat offhand way of defining the XML document root,
which happened to work well with the old MXML versions, maybe rather by accident:
It appended new Element nodes (tags), but set their element name to the content
of the Directive ("?xml...") or Definition ("!DOCTYPE...").

This solution breaks on the new MXML v4, for several reasons
 * MXML seems to ensure now that actual tags/elements are always properly closed
 * MXML now XML-escapes the names of all elements.

Furthermore, there was already a dedicated function mxmlNewXML() in v3,
and there are now dedicated functions for Directrives and Definitions,
together with special type constants to detect those kinds of nodes.

Basically the solution follows the established scheme:
we delegate to a forked low-level implementation in the Policy baseclass...

However, doing so creates a conflict in the design:
- we do not want XMLtree to become a generic XML builder -- it is dedicated to Zyn/Yoshimi format solely
- but we also do not want a mutual dependency of XMLStore and XMLtree at API level;
  there is in fact a dependency, but only since both depend on MXML implementation details.

I solved this dilemma by introducing a public configuration function on XMLtree,
which however uses an /opaque enum/ as argument, and can thus, technically,
only be invoked from within XMLStore.cpp, where the shared back-end resides.

Furthermore, I reformulated the Whitespace-callback,
both to make it more readable and to allow using it with both MXML versions

__RESULT__: with this change, the XML files can be reproduced, with minor differences:
- the encoding "utf-8" is now lowercase
- trailing whitespace in the attribute list of tags is removed.
- there is no trailing newline.
All these difference are caused by the new (much improved) MXML implementation
and can thus be considered unavoidable.